% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{aws.ml-package}
\alias{aws.ml-package}
\alias{aws.ml}
\title{aws.ml}
\description{
Client for the Amazon Machine Learning
}
\details{
Client for the Amazon Machine Learning
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
