% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_source.R
\name{create_data_source}
\alias{create_data_source}
\alias{delete_data_source}
\alias{spec_s3}
\alias{spec_rds}
\alias{spec_redshift}
\title{Create Data Source(s)}
\usage{
create_data_source(id, specification, name = NULL,
  compute_statistics = TRUE, iam_role = NULL, ...)

delete_data_source(id, ...)

spec_s3(data, schema, rearrangement)

spec_rds(data, schema, rearrangement)

spec_redshift(data, schema, rearrangement)
}
\arguments{
\item{id}{A character string specifying an identifier for the model}

\item{specification}{A list.}

\item{name}{Optionally, a character string specifying a name for the data source}

\item{compute_statistics}{A logical. This should be TRUE.}

\item{iam_role}{If \code{specification} is for an RDS or Redshift source, an IAM role ARN.}
}
\description{
Create a data source for an AWS ML Model
}
