% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_sources.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List Data Sources}
\usage{
list_data_sources(n = 100, token = NULL, ...)
}
\arguments{
\item{n}{An integer specifying the maximum number of results to return.}

\item{token}{A character string specifying a pagination token.}
}
\description{
Return a list of AWS data sources
}
\references{
\urL{https://docs.aws.amazon.com/machine-learning/latest/APIReference/API_DescribeDataSources.html}
}
