% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_models.R
\name{list_models}
\alias{list_models}
\title{List Models}
\usage{
list_models(n = 100, token = NULL, ...)
}
\arguments{
\item{n}{An integer specifying the maximum number of results to return.}

\item{token}{A character string specifying a pagination token.}
}
\description{
Return a list of AWS ML Models
}
\references{
\urL{https://docs.aws.amazon.com/machine-learning/latest/APIReference/API_DescribeMLModels.html}
}
